Introduce rov_utilities. Remove vortex_msg
rov_utilities contains rov specific functions, such as updating pwm signals,
arm/disarm/start sequences and so on. Additionally, it contains definitions such as
number of thrusters and vortex_msg. The idea is to let the "hardware related" functionality
be in their own files, and rov functions in this one.

TODO in the future:
- Expand to several files when making AUV
- Sensor fusion may be done on MCU
- Autonomy related tasks
- Expand this list