Add new branch: efm32gg990f1024
The meaning of this branch is to provide a program that works on the STK3700
starter kit with the efm32gg990f1024 chip. On the PCB on the ROV, the MCU is
named EFM32GG942F1024.

The differences between the two MCUs is unclear, but it is clear that we can't
load the same binary files in both of them.