Fixes #10
An additional check is made in USART1_RX_IRQHandler() to handle that MAGIC_STOP_BYTE comes before it is supposed to. This will
fix the issue for now, but in the future, the IRQHandler should do
"less". Meaning it should only set a flag or signal a semaphore
before exiting.