Add thruster mapping enum. Add light pwm functionality
The thruster mapping is done via the thruster_mapping enumeration.
For now, the mapping is the same as before. Before remapping as it is
supposed to be, the mapping should be removed in mcu_interface first.

Additionally, the boundary check on PWM data is moved out from
us_to_comparevalue() to the pwm_update() function.